Search results now work with tokens and a cache.
When doing a regex based search the MessageStore returns a token which
is a unique identifier for that search. The status of the search can be
checked with the token to see if results are ready or not. If results
are ready then they can be retrieved with the token as well. All cached
search results have a configurable TTL which defaults to 24hrs.